+++ slim/base/user.py
@@ -1,5 +1,9 @@
+import typing
 from abc import abstractmethod
-from typing import Type
+from typing import Union
+
+if typing.TYPE_CHECKING:
+    from .view import BaseView, Type
 
 
 class BaseUser:
@@ -24,65 +28,70 @@ class BaseUserViewMixin:
 
     from model.user import User
 
-    class UserViewMixin(BaseUserMixin):
+    class UserViewMixin(BaseUserViewMixin):
         @property
         def user_cls(self) -> Type[BaseUser]:
             return User
 
     """
 
+    def get_current_user(self: Union['BaseUserViewMixin', 'BaseView']):
+        key = self.get_user_token()
+        if key: return self.get_user_by_token(key)
+
+    @property
     @abstractmethod
-    def get_current_user(self):
-        """Override to determine the current user from, e.g., a cookie."""
+    def user_cls(self) -> Type[BaseUser]:
         pass
 
     @abstractmethod
-    def get_user_by_key(self, key):
+    def get_user_token(self: Union['BaseUserViewMixin', 'BaseView']):
+        """Get access token for specified user"""
         pass
 
     @abstractmethod
-    def setup_user_key(self, key, expires=30):
-        """ setup user key for server """
+    def get_user_by_token(self: Union['BaseUserViewMixin', 'BaseView'], token) -> Type[BaseUser]:
         pass
 
     @abstractmethod
-    def teardown_user_key(self):
-        """ teardown user key for server, make the token invalid here"""
+    def setup_user_token(self: Union['BaseUserViewMixin', 'BaseView'], user_id, key=None, expires=30):
+        """ setup user token """
+        pass
+
+    @abstractmethod
+    def teardown_user_token(self: Union['BaseUserViewMixin', 'BaseView'], token=None):
+        """ invalidate the token here"""
         pass
 
 
 # noinspection PyAbstractClass
 class BaseSecureCookieUserViewMixin(BaseUserViewMixin):
-    def get_current_user(self):
+    def get_user_token(self: Union[BaseUserViewMixin, 'BaseView']):
+        """Get access token for specified user"""
         try:
-            key = self.get_secure_cookie('u')
-            if key:
-                return self.get_user_by_key(key)
+            return self.get_secure_cookie('u')
         except:
             self.del_cookie('u')
 
-    def setup_user_key(self, key, expires=30):
-        self.set_secure_cookie('u', key, max_age=expires)
+    def setup_user_token(self: Union[BaseUserViewMixin, 'BaseView'], user_id, key=None, expires=30):
+        """ setup user token """
+        if key:
+            self.set_secure_cookie('u', key, max_age=expires)
 
-    def teardown_user_key(self):
+    def teardown_user_token(self: Union['BaseUserViewMixin', 'BaseView'], token=None):
+        """ invalidate the token here"""
         self.del_cookie('u')
 
 
 # noinspection PyAbstractClass
 class BaseAccessTokenUserViewMixin(BaseUserViewMixin):
-    def get_current_user(self):
-        access_token = self.headers.get('AccessToken', None)
-        if access_token: return self.get_user_by_key(access_token)
-
-    def setup_user_key(self, key, expires=30):
-        pass
+    def get_user_token(self: Union[BaseUserViewMixin, 'BaseView']):
+        """Get access token for specified user"""
+        return self.headers.get('AccessToken', None)
 
 
 # noinspection PyAbstractClass
 class BaseAccessTokenInParamUserViewMixin(BaseUserViewMixin):
-    def get_current_user(self):
-        access_token = self.params.get('AccessToken', None)
-        if access_token: return self.get_user_by_key(access_token)
-
-    def setup_user_key(self, key, expires=30):
-        pass
+    def get_user_token(self: Union[BaseUserViewMixin, 'BaseView']):
+        """Get access token for specified user"""
+        return self.params.get('AccessToken', None)